pyproject.toml
[tool.poetry]
name = "sistemas_gestao_projetos"
version = "0.1.0"
description = "Sistema para Gestao de Projetos de Engenharia de Software"
authors = ["Willames <williames@arbache.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.3"
djangorestframework = "^3.13.1"
python-decouple = "^3.6"
psycopg2 = "^2.9.3"
django-filter = "^21.1"
celery = "^5.2.6"
redis = "^4.3.1"
django-cors-headers = "^3.12.0"
whitenoise = "^6.1.0"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
ipython = "^8.2.0"
flake8 = "^4.0.1"
factory-boy = "^3.2.1"
pre-commit = "^2.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
